% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitiMoMo.R
\name{bootstrap.fitiMoMo}
\alias{bootstrap.fitiMoMo}
\title{Bootstrap an Improvement Rate Mortality Model}
\usage{
\method{bootstrap}{fitiMoMo}(
  object,
  nBoot = 1,
  type = c("semiparametric", "residual"),
  deathType = c("observed", "fitted"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"fitiMoMo"} with the fitted
parameters of an improvement rate model.}

\item{nBoot}{number of bootstrap samples to produce.}

\item{type}{type of bootstrapping approach to be applied.
\code{"semiparametric"}(default) uses the assumed distribution of the
deaths to generate bootstrap samples. \code{"residual"} resamples the
deviance residuals of the model to generate bootstrap samples.}

\item{deathType}{type of deaths to sample in the semiparametric bootstrap.
\code{"observed"} (default) resamples the observed deaths. \code{"fitted"}
resamples the fitted deaths. This parameter is only used if \code{type} is
\code{"semiparametric"}.}
}
\description{
Produce bootstrap parameters of an improvement rate model to account for
parameter uncertainty.
}
