% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitiMoMo.R
\name{bootstrap.fitiMoMoCIF}
\alias{bootstrap.fitiMoMoCIF}
\title{Bootstrap an Improvement Rate Mortality Model of type fitted
with constant improvement rates}
\usage{
\method{bootstrap}{fitiMoMoCIF}(object, nBoot = 1,
  type = c("semiparametric", "residual"), deathType = c("observed",
  "fitted"), ...)
}
\description{
Bootstrap an Improvement Rate Mortality Model of type fitted
with constant improvement rates
}

