% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitiMoMo.R
\name{bootstrap.fitiMoMoF}
\alias{bootstrap.fitiMoMoF}
\title{Bootstrap an Improvement Rate Mortality Model of type fitted}
\usage{
\method{bootstrap}{fitiMoMoF}(object, nBoot = 1,
  type = c("semiparametric", "residual"), deathType = c("observed",
  "fitted"), ...)
}
\description{
Bootstrap an Improvement Rate Mortality Model of type fitted
}
