% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitiMoMo.R
\name{bootstrap.fitiMoMoO}
\alias{bootstrap.fitiMoMoO}
\title{Bootstrap an Improvement Rate Mortality Model of type observed}
\usage{
\method{bootstrap}{fitiMoMoO}(object, nBoot = 1, type = c("semiparametric",
  "residual"), deathType = c("observed", "fitted"), ...)
}
\description{
Bootstrap an Improvement Rate Mortality Model of type observed
}

