% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardModels.R
\name{cbdi}
\alias{cbdi}
\title{Create a Cairns-Blake-Dowd model for improvement rates}
\usage{
cbdi(type = c("indirect", "direct"))
}
\arguments{
\item{type}{defines the type of estimation method to be used.
\code{"indirect"} would estimate the equivalent mortality rate model
and the transform the model into improvement rates. \code{"direct"}
would estimate the model directly on the improvement rate model.}
}
\description{
Create a Cairns-Blake-Dowd model for improvement rates
}
