% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitiMoMoD.R
\name{fit.iMoMoD}
\alias{fit.iMoMoD}
\title{Fit an Improvement Rate Model  directly}
\usage{
\method{fit}{iMoMoD}(object, data = NULL, Dxt = NULL, Ext = NULL,
  ages = NULL, years = NULL, ages.fit = NULL, years.fit = NULL,
  wxt = NULL, start.ax = NULL, start.bx = NULL, start.kt = NULL,
  start.b0x = NULL, start.gc = NULL, verbose = TRUE, ...)
}
\description{
Fit an Improvement Rate Model directly to the improvement rates
}
