% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitiMoMoI.R
\name{fit.iMoMoI}
\alias{fit.iMoMoI}
\title{Fit an Improvement Rate Model indirectly}
\usage{
\method{fit}{iMoMoI}(
  object,
  data = NULL,
  Dxt = NULL,
  Ext = NULL,
  ages = NULL,
  years = NULL,
  ages.fit = NULL,
  years.fit = NULL,
  wxt = NULL,
  start.ax = NULL,
  start.bx = NULL,
  start.kt = NULL,
  start.b0x = NULL,
  start.gc = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{"iMoMoI"} defining the improvement rate
model.}

\item{data}{an optional object of type StMoMoData containing information on
deaths and exposures to be used for fitting the model. This is typically created
with  function \code{\link{StMoMoData}}. If this is not provided then the fitting
data is taken from arguments, \code{Dxt}, \code{Ext}, \code{ages}, \code{years}.}

\item{Dxt}{optional matrix of deaths data.}

\item{Ext}{optional matrix of observed exposures of the same dimension of
\code{Dxt}.}

\item{ages}{optional vector of ages corresponding to rows of \code{Dxt} and
\code{Ext}.}

\item{years}{optional vector of years corresponding to rows of \code{Dxt} and
\code{Ext}.}

\item{ages.fit}{optional vector of ages to include in the fit. Must be a
subset of \code{ages}.}

\item{years.fit}{optional vector of years to include in the fit. Must be a
subset of \code{years}.}

\item{wxt}{optional matrix of 0-1 weights to be used in the fitting process.
This can be used, for instance, to zero weight some cohorts in the data.
See \code{\link{genWeightMat}} which is a helper function for defining
weighting matrices.}

\item{start.ax}{optional vector with starting values for \eqn{\alpha_x}.}

\item{start.bx}{optional matrix with starting values for \eqn{\beta_x^{(i)}}.}

\item{start.kt}{optional matrix with starting values for \eqn{\kappa_t^{(i)}}.}

\item{start.b0x}{optional vector with starting values for \eqn{\beta_x^{(0)}}.}

\item{start.gc}{optional vector with starting values for \eqn{\gamma_c}.}

\item{verbose}{a logical value. If \code{TRUE} progress indicators are
printed as the model is fitted. Set \code{verbose = FALSE} to silent the
fitting and avoid progress messages.}

\item{...}{arguments to be passed to or from other methods. This can be
used to control the fitting parameters of \code{gnm}. See
\code{\link[gnm]{gnm}}.}
}
\description{
Fit an Improvement Rate Model indirectly via fiting the equivalent mortality
rate model and then trasforming the output to improvement rates
}
