% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleVAR.r
\name{forecast.simpleVAR}
\alias{forecast.simpleVAR}
\title{Forecast a VAR from class simpleVAR}
\usage{
\method{forecast}{simpleVAR}(object, h = 10, colnames = NULL)
}
\description{
ARGUMENT
}
\details{
object: An object of class "ets", "Arima" or "ar".
n.ahead: Number of periods for the simulated series
colnames: Name of the columns (time periods)
}

