% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfitiMoMo.R
\name{getMinimalFitiMoMo}
\alias{getMinimalFitiMoMo}
\title{Extract a lighter version of a  Improvment Rate Mortality Model}
\usage{
getMinimalFitiMoMo(object)
}
\description{
Extract a lighter version of a  Improvment Rate Mortality Model
}
\keyword{internal}
