% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dociMoMo.R, R/iMoMo.R
\docType{package}
\name{iMoMo}
\alias{iMoMo}
\alias{iMoMo-package}
\title{iMoMo.}
\usage{
iMoMo(model, type = c("observed", "fitted"))
}
\description{
iMoMo.

Create a new Imprvement Rate Mortality Model
}

