% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardModels.R
\name{lci}
\alias{lci}
\title{Create a Lee-Carter model for improvment rates}
\usage{
lci(type = c("indirect", "direct"), const = c("sum", "last", "first"))
}
\arguments{
\item{type}{defines the type of estimation method to be used.
\code{"indirect"} would estimate the equivalent mortality rate model
and the transform the model into improvement rates. \code{"direct"}
would estimate the model directly on the improvement rate model.}

\item{const}{defines the constraint to impose to the period index of the
model to ensure identifiability. The alternatives are
\code{"sum"}(default),  \code{"last"} and \code{"first"} which apply
constraints \eqn{\sum_t\kappa_t = 0}, \eqn{\kappa_n = 0} and
\eqn{\kappa_1 = 0}, respectively.}
}
\value{
An object of class \code{"StMoMo"}.
}
\description{
Utility function to initialise a \code{iMoMo} object representing a
Lee-Carter model on improvment rates.
}
\details{
The created model  has predictor structure
\deqn{\log(\mu_{xt}/\mu_{x,t-1}) = \alpha_x + \beta_x\kappa_t.}
To ensure identifiability one of the  following constraints is imposed
\deqn{\sum_t\kappa_t = 0,\,\kappa_1 = 0,\, \kappa_n = 0}
depending on the value of \code{const}, and
\deqn{\sum_x\beta_x = X.}
where \eqn{X}  is the number of ages in the data
}
\seealso{
\code{\link{iMoMo}}
}
