% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTwoBootstraps.R
\name{plotParameterFan}
\alias{plotParameterFan}
\title{Plot fanchart of the parameters}
\usage{
plotParameterFan(x, y, yBoot, main, xlab, ylab, probs, fan.col, n.fan,
  new = TRUE, ...)
}
\description{
Plot fanchart of the parameters
}
\keyword{internal}

