% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotTwoModels.R
\name{plotTwoModels}
\alias{plotTwoModels}
\title{Plot fitted parameters from two stochastic mortality model}
\usage{
plotTwoModels(x, y, nCol = 2, parametricbx = TRUE, type = "l",
  legend = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"fitStMoMo"} with the fitted
parameters of a stochastic mortality model.}

\item{y}{an object of class \code{"fitStMoMo"} with the fitted
parameters of a stochastic mortality model.}

\item{nCol}{number of columns to use in the plot.}

\item{parametricbx}{if \code{FALSE} parametric age-modulating terms,
which don't need to be estimated, are not plotted.}

\item{type}{what type of plot should be drawn. See
\code{\link[graphics]{plot}}.}

\item{...}{additional arguments to control graphical appearance.
See \code{\link[graphics]{plot}}.}
}
\description{
Plot fitted parameters from two stochastic mortality model
}

