% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualsfitiMoMo.R
\name{residuals.fitiMoMoI}
\alias{residuals.fitiMoMoI}
\title{Extract deviance residuals of an Improvement Rate  Model}
\usage{
\method{residuals}{fitiMoMoI}(object, scale = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{"fitiMoMo"} with the fitted
parameters of an improvment mortality model.}

\item{scale}{logical indicating whether the residuals should be scaled
or not by dividing the deviance by the  overdispersion of the model.
Default is \code{TRUE}.}

\item{...}{other arguments.}
}
\value{
An object of class \code{"resStMoMo"} with the residuals. This
object has components:
  \item{residuals}{ a matrix with the residuals.}
  \item{ages}{ ages corresponding to the rows in \code{residuals}.}
  \item{years}{ years corresponding to the columns in \code{residuals}.}
}
\description{
Compute deviance residuals of a fitted improvement rate  Model.
These residuals can be plotted using \code{\link[StMoMo]{plot.resStMoMo}}.
}
\seealso{
\code{\link[StMoMo]{plot.resStMoMo}}
}
