% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StMoMo2iMoMo.R
\name{StMoMo2iMoMo}
\alias{StMoMo2iMoMo}
\title{Transform a mortality rate model to an improvement rate model}
\usage{
StMoMo2iMoMo(model, type = c("indirect", "direct"))
}
\arguments{
\item{model}{A StMoMo object representing the mortality rate model
to be transformed into an improvement rate model. This StMoMo
model must include an static age function \eqn{\alpha_x} and must
use a \code{"log"} link.}

\item{type}{defines the type of estimation method to be used.
\code{"indirect"} would estimate the equivalent mortality rate model
and the transform the model into improvement rates. \code{"direct"}
would estimate the model directly on the improvement rate model.}
}
\description{
Transforms a StMoMo object representing a mortality rate model
into an iMoMo object representing an improvement rate model
}
