% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoCIF}
\alias{transParamiMoMoCIF}
\title{Transform fitted parameters of fitted improvement
rated model into the appropriate output format}
\usage{
transParamiMoMoCIF(out)
}
\description{
Transform fitted parameters of fitted improvement
rated model into the appropriate output format
}
\keyword{internal}
