% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoD}
\alias{transParamiMoMoD}
\title{Transform fitted parameters of direct improvement
rated model into the appropriate output format}
\usage{
transParamiMoMoD(out)
}
\description{
Transform fitted parameters of direct improvement
rated model into the appropriate output format
}
\keyword{internal}
