% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoF}
\alias{transParamiMoMoF}
\title{Transform fitted parameters of fitted improvement
rated model into the appropriate output format}
\usage{
transParamiMoMoF(out)
}
\description{
Transform fitted parameters of fitted improvement
rated model into the appropriate output format
}
\keyword{internal}
