% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoI}
\alias{transParamiMoMoI}
\title{Transform fitted parameters of indirect improvement
rated model without constanst improvements into the
appropriate output format}
\usage{
transParamiMoMoI(out)
}
\description{
Transform fitted parameters of indirect improvement
rated model without constanst improvements into the
appropriate output format
}
\keyword{internal}
