% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoI_CI}
\alias{transParamiMoMoI_CI}
\title{Transform fitted parameters of indirect improvement
rated model with constant improvement rates into
the appropriate output format}
\usage{
transParamiMoMoI_CI(out, constFunEst)
}
\description{
Transform fitted parameters of indirect improvement
rated model with constant improvement rates into
the appropriate output format
}
\keyword{internal}
