% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalUtils.R
\name{transParamiMoMoO}
\alias{transParamiMoMoO}
\title{Transform fitted parameters of observed improvement
rated model into the appropriate output format}
\usage{
transParamiMoMoO(out)
}
\description{
Transform fitted parameters of observed improvement
rated model into the appropriate output format
}
\keyword{internal}

