% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedfitiMoMo.R
\name{fitted.fitiMoMo}
\alias{fitted.fitiMoMo}
\title{Compute fitted values for a Stochastic Mortality Model}
\usage{
\method{fitted}{fitiMoMo}(object, type = c("improvements", "rates",
  "deaths"), ...)
}
\description{
Returns fitted values for the data used in Improvement Rate Stochastic
Mortality Model.
}
