% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleVAR.r
\name{forecast.simpleVAR}
\alias{forecast.simpleVAR}
\title{Forecast a VAR from class simpleVAR}
\usage{
\method{forecast}{simpleVAR}(object, h = 10, colnames = NULL)
}
\arguments{
\item{object}{An object of class \code{"simpleVAR"}.}

\item{h}{Number of periods for the simulated series}

\item{colnames}{Name of the columns (time periods)}
}
\value{
A matrix with the forecast series
}
\description{
Forecast a VAR from class simpleVAR
}
