% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleVAR2.r
\name{simpleVAR2}
\alias{simpleVAR2}
\title{Simple VAR estimation}
\usage{
simpleVAR2(y, p = 1, d = 0, type = c("const", "trend", "both", "none"))
}
\description{
y: a matrix containing the multivariate timeseries
p: Integer for the lag order (default is p=1)
d: Degree of differencing (default is d=0)
}

