% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatebootiMoMo.R
\name{simulate.bootiMoMo}
\alias{simulate.bootiMoMo}
\title{Simulate future sample paths from a Bootstrapped Improvement
Rate Model}
\usage{
\method{simulate}{bootiMoMo}(
  object,
  nsim = 1,
  seed = NULL,
  h = 50,
  kt.order = c(1, 0, 0),
  kt.include.constant = TRUE,
  kt.include.trend = FALSE,
  gc.order = c(1, 0, 0),
  gc.include.constant = TRUE,
  jumpRates = NULL,
  kt.lookback = NULL,
  gc.lookback = NULL,
  ...
)
}
\description{
Simulate future sample paths from a Bootstrapped Improvement
Rate Model
}
