% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatefitiMoMo.R
\name{simulate.fitiMoMo}
\alias{simulate.fitiMoMo}
\title{Simulate future sample paths from an improvement  Model}
\usage{
\method{simulate}{fitiMoMo}(object, nsim = 1000, seed = NULL, h = 50,
  kt.order = c(1, 0, 0), kt.include.constant = TRUE,
  kt.include.trend = FALSE, gc.order = c(1, 1, 0),
  gc.include.constant = TRUE, jumpRates = NULL, kt.lookback = NULL,
  gc.lookback = NULL, ...)
}
\description{
Simulate future sample paths from an improvement  Model
}

