% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleVAR.r
\name{simulate.simpleVAR}
\alias{simulate.simpleVAR}
\title{Simulate n.path of VAR from class simpleVAR}
\usage{
\method{simulate}{simpleVAR}(object, nsim = 10, seed = NULL,
  colnames = NULL)
}
\description{
ARGUMENT
}
\details{
object: An object of class "ets", "Arima" or "ar".
n.ahead: Number of periods for the simulated series
n.path: Number of paths to be simulated
colnames: Name of the columns (time periods)
}

